Updated: content create method
