+++ Routes/backend/routes-contents.php
@@ -1,10 +1,24 @@
 <?php
 
+
+Route::group(
+    [
+        'prefix' => 'backend/cms/media',
+        'namespace'  => 'Backend',
+        'middleware' => ['web', 'has.backend.access'],
+    ],
+    function () {
+        //---------------------------------------------------------
+        Route::post('/upload', 'MediaController@upload')
+            ->name('backend.cms.media.upload');
+        //---------------------------------------------------------
+    });
+
 Route::group(
     [
         'prefix' => 'backend/cms/contents/{content_type_slug}',
 
-        'middleware' => ['web', 'has.backend.access'],
+        'middleware' => ['web', 'has.backend.access', 'set.content.type'],
 
         'namespace' => 'Backend',
 ],
